Segundo Commit

+++ src/pages/Repositories/index.js
@@ -1,14 +1,22 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useNavigate } from 'react-router-dom';
+
 /*useEffect + useState ok */
 export default function Repositories() {
+    const history = useNavigate()
     const [ repositories, setRepositories ] = useState([]);
 
     useEffect(() =>{
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if(repositoriesName != null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            console.log("aaaaaaaa")
+            history('/');  //useHistory está outdated, agr usamos o useNavigate > não precisa do .push
+        }
     }, []);
     return (
         <S.Container>
@@ -20,6 +28,7 @@ export default function Repositories() {
                     )
                 })}
             </S.List>
+            <S.LinkHome to='/'>Voltar</S.LinkHome>
         </S.Container>
     )
 }
